<std::string, std::vector<float>>>
 ReadCsv_string_int_pair(std::string filename);

+++ EPOS_Linux_Library/examples/HelloEposCmd/profile_velocity_mode_CAN.cpp
@@ -108,7 +108,7 @@ void  PDO_Mapping(unsigned int *p_pErrorCode,unsigned short g_usNodeId_local);
 int   PrepareCyclicTorqueMode(unsigned int* p_pErrorCode,unsigned short g_usNodeId_local);
 int  CyclicSynchronusTroqueModeSettings(HANDLE p_DeviceHandle, unsigned short p_usNodeId , unsigned int * p_rlErrorCode, int lResult);
 int  ProfileVelocityModeSettings(HANDLE p_DeviceHandle, unsigned short p_usNodeId , unsigned int *p_rlErrorCode,int lResult);
-std::vector<std::pair<std::string, std::vector<int>>> ReadCsv_string_int_pair(std::string filename);
+std::vector<std::pair<std::string, std::vector<float>>> ReadCsv_string_int_pair(std::string filename);
 
 std::vector<std::pair<std::string, std::vector<float>>> ReadCsv_string_int_pair(std::string filename)
 {
